\name{DescargarLibros}
\alias{DescargarLibros}
\title{DescargarLibros}
\usage{
DescargarLibros(carpeta)
}
\arguments{
  \item{carpeta}{Character. Especifica la ruta de la carpeta en la cual se desean descargar los libros}
}
\description{
Una función que descarga los libros liberados por Springer ante la contingencia COVID-19
}
\examples{

## Not run: DescargarLibros(carpeta)
}
